Fix/crash open recent project 1122 (#1138)
* fix: prevent crashes when opening project

* fix: generalize error message for non-.inv3 file selection

* fix: simplify ResetTrackerFiducials and fix same dtype issue in image.py

* style: remove extra blank line in tracker.py (ruff-format)

+++ invesalius/navigation/image.py
@@ -56,7 +56,7 @@ class Image:
         session = ses.Session()
         state = session.GetState("image")
         if state is not None:
-            self.fiducials = np.array(state["image_fiducials"])
+            self.fiducials = np.array(state["image_fiducials"], dtype=np.float64)
 
     def LoadProject(self) -> None:
         self.fiducials = prj.Project().image_fiducials
